requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
aiohttp==3.9.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
markdown==3.5.2
pydantic==2.5.2
certifi==2023.11.17
urllib3==2.1.0
lxml==4.9.3 